requirements.txt
accelerate==0.24.1
torchvision
transformers>=4.25.1
ftfy
tensorboard
Jinja2
git+https://github.com/huggingface/diffusers